requirements.txt
# core
jax[cuda12_pip]
mujoco_mjx
brax
torch
numpy 
gymnasium
playground
torchaudio

# utils
matplotlib
py-lz4framed
hydra-core
jupyter
ipython
ipykernel
wandb
mediapy
dill